Add time and realtime clock functions
Adds functions to query the system's realtime clock, convert time intervals to/from a calendar date and time in either UTC or the local time, and format dates and times as strings.

Two structs, SDL_TimeSpec (a time interval represented in seconds + nanoseconds) and SDL_DateTime (broken down calendar date and time), were added to facilitate this functionality.

Querying the system time results in a value expressed in seconds and nanoseconds since the Unix epoch (Jan 1, 1970) in UTC +0000. Conversions to and from the various platform epochs and units are performed when required.

Any direct handling of timezones and DST were intentionally avoided, as different platforms have different limitations on providing this information accurately, particularly when dealing with dates from the past, and some platforms can only provide whether DST is currently active as opposed to anything historical. The offset from UTC is provided when converting from UTC to a local time by calculating the difference between the original UTC and the resulting local time, and the DST status of a point in time is forwarded to the application if the platform reports it, but otherwise timezone and DST info and calculations are completely left to the underlying platform.

Formatting of date and time strings is handled by the SDL_FormatDateTime() function, which mimics the functionality of strftime, but with better consistency across platforms, even if the underlying platform's libc lacks or provides a limited implementation of strftime. In cases where it is available, the underlying strftime is used when encountering certain format codes to provide the preferred strings and date/time formats according to the system locale. Some highly locale dependent functionality, such as the E and O modifiers for handling eras and alternate numerical formats, is only available if the underlying platform has the proper support.

Helper functions for calculating the difference between SDL_TimeSpec structures are included.

An automated test for the time conversion and string formatting function was added, as is a simple standalone test to display the current system date and time onscreen.