EGL: Fix transparent windows on certain Nvidia drivers
Some Nvidia EGL implementations ignore the value parameter for the EGL_EXT_present_opaque extension and enable it based solely on the presence of the GL_PRESENT_OPAQUE_EXT enum in the attribute array. Only request this extension if the intent is to enable it, otherwise omit it.

This fixes transparent windows on Nvidia, and has no behavioral changes on other drivers, since not requesting this extension is the same as setting it to EGL_FALSE.