wayland: Implement threaded input handling
Create a separate event queue for keyboard and mouse input and run it on its own thread. This cuts input latency down significantly by delivering the input events as fast as the system can forward them, typically reducing latency from several milliseconds to less than 1 millisecond on average, particularly for client applications using an event filter to get events as soon as they are added to the queue.