win32: Detect tiled windows
Windows provides the IsWindowArranged() function, which must be manually loaded from User32.dll, for detecting if a window is in the tiled state.

Load and use it, if available, to detect the tiled state and prevent overwriting the non-tiled size and position with the size and position sent when the window is in tiled mode.