core/unix: Add support for ppoll() and remove the select() fallback path
Add support for ppoll() to allow nanosecond timing precision, support for polling multiple fds, and remove the select() path in favor of poll() being the standard fallback in teh absence of ppoll().

poll() has been available since at least AT&T Sys V release 3 in 1987, OpenBSD 2.2 in 1994, NetBSD 1.3 and FreeBSD 3.0 in 1996, and Linux 2.1.23 in 1997, so there is no danger of building SDL3 on a platform that doesn't support it.