wayland: Disable the EGL frame callback hack if FIFO forward progress is guaranteed
The eglSwapBuffers function may block indefinitely if frame callbacks stop, which necessitated manual handling of the callback and presentation to ensure that forward progress was always possible, however, if the required protocols for guaranteeing FIFO forward progress are present, this is no longer required.

Disable the frame callback hack and use the standard swap interval and presentation functions if the fifo_v1 and commit_timing_v1 protocols are supported, and the low-latency double buffer hint is not enabled.