wayland/x11: Add a function for retrieving the xkb keysyms for a raw keycode
Currently, the raw key codes as returned by the X11 and Wayland backends during key events are fairly useless, as they don't correspond to any functionality in particular without also knowing the current keyboard mapping, which is internal on X11, and would require exposing the mapped memory on Wayland, on top of requiring clients to pull in the X11/xkbcommon libraries themselves.

This adds a simple function for retrieving the XKB keysym value for the raw keycode, so that the intended function of esoteric keys can be easily looked up, in the event that SDL has no known mapping and returns an 'SDL_SCANCODE_UNKNOWN' or 'SDLK_UNKNOWN' value.