events: Add reuse and fast switching of keymaps
Wayland multi-seat can have multiple keyboards with independent active layouts, which require fast keymap switching to multiplex to the global keyboard device. Add a flag to allow keymap lifetimes to be externally managed, and store the layout info with the keymap. This reduces the overhead of switching keymaps to setting a single pointer.