video: Handle overlapping display rects in the display selection algorithm
In the case of the centerpoint of the window being enclosed within multiple overlapping display rects, which can occur with certain scaled desktop layouts, favor the bottommost/rightmost display, as that is the most likely candidate for the 'true' window position.

In the case where no displays are overlapping, there is no change from the previous behavior.