wayland: Improve pointer confinement reliability
If the pointer should be confined, keep trying until a confine/lock signal is received. This considerably improves reliability on compositors where confining can be a racy operation.

A pointer lock is used to special-case 1x1 confinement regions, as otherwise, the pointer can still move at the subpixel level, particularly on scaled desktops.