wayland: Refactor keyboard layout handling
This adds a small amount overhead to the initial keymap building process, which is already significantly reduced in previous commits, reducing the cost of switching layouts to just swapping a pointer if the keymap doesn't actually have to be rebuilt, which is often doesn't when switching between layouts with common character sets.

Additionally, handling of unknown keysyms, particularly when dealing with virtual keyboards, is much improved, as keys generating valid Unicode values with no corresponding scancode will be dynamically added to the keymap with reserved scancodes.